fixed: A couple of minor warnings.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@17409 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/karaoke/karaokelyricstextkar.cpp
@@ -469,7 +469,7 @@ void CKaraokeLyricsTextKAR::parseMIDI()
     if ( (int) lyrics[i].track != preferred_lyrics_track )
       continue;
 
-    unsigned int mstime = ceil( (mts.advanceClocks( lyrics[i].clocks ) - firstNoteTime) / 100);
+    unsigned int mstime = (unsigned int)ceil( (mts.advanceClocks( lyrics[i].clocks ) - firstNoteTime) / 100);
     addLyrics( lyrics[i].text, mstime, lyrics[i].flags );
   }
 }
